+++ tools/configure_deps.R
@@ -9,12 +9,13 @@ Sys.setenv(
 )
 
 library(reticulate)
-py_require(c(
-  "markitdown[all]",
-  if (identical(.Platform$OS.type, "windows")) {
-    py_require("onnxruntime<=1.20.1")
-  }
-))
+py_require(
+  c(
+    "markitdown[all]",
+    if (identical(.Platform$OS.type, "windows")) "onnxruntime<=1.20.1"
+  ),
+  python_version = "<3.14"
+)
 try({
   print(py_config())
   import("markitdown")